Add initial output for Lambda flare command
