+++ src/commands/lambda/cli.ts
@@ -1,4 +1,5 @@
+import {LambdaFlareCommand} from './flare'
 import {InstrumentCommand} from './instrument'
 import {UninstrumentCommand} from './uninstrument'
 
-module.exports = [InstrumentCommand, UninstrumentCommand]
+module.exports = [InstrumentCommand, UninstrumentCommand, LambdaFlareCommand]
